Update 42_redelk-base-docker-init.sh
Since user `redelk` is the only one creating files in `/var/www/html/c2logs`, it should be the owner, and `www-data` as the group owner so that nginx can serve the files. 

SGID is set for  `/var/www/html/c2logs` so that the www-data group owner will be propagated for all files created through `getremotelogs.sh` cron job, so that  there is no need to a chmod/chown after the cron job to fix permission issue stopping nginx from serving the log files.